Python-Tutorials/streamlit/requirements.txt
imutils==0.5.3
numpy==1.17.4
streamlit==0.80.0
Pillow==8.3.0
opencv-python-headless